Finalize Python Environment API
